+++ previous-versions/banking/src/sdk/models/operations/getaccount.ts
@@ -55,5 +55,5 @@ export class GetAccountResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }