+++ previous-versions/banking/src/sdk/models/operations/gettransactioncategory.ts
@@ -49,7 +49,7 @@ export class GetTransactionCategoryResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Success