+++ previous-versions/banking/src/sdk/models/operations/listaccounts.ts
@@ -73,5 +73,5 @@ export class ListAccountsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }