+++ previous-versions/banking/src/sdk/models/operations/listtransactions.ts
@@ -67,7 +67,7 @@ export class ListTransactionsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Success